change decimal to double in Quantity and Price

+++ CRM/CRM/Models/Binance/AccountTradeHistory.cs
@@ -10,8 +10,8 @@ namespace CRM.Models.Binance
         public DateTime Time { get; set; }
         public string Side { get; set; }
         public string Pair { get; set; }
-        public decimal Price { get; set; }
-        public decimal Quantity { get; set; }
+        public double Price { get; set; }
+        public double Quantity { get; set; }
         public decimal DollarQuantity { get; set; }
         public double BalanceUSDT { get; set; }
         public double Profit { get; set; }
